+++ src/client.cjs
@@ -1,5 +1,5 @@
 window.__ModuleLoader__.load({
-  id: 'dsh-minimal-first-turn',
+  id: '@feiyueve/dsh-minimal-first-turn',
   factory: (require) => {
     var module = { exports: {} }
     var exports = module.exports